rename: model.py to mcdropout.py
